Fixed some bugs, added tree creator

+++ src/DBMSC_defs.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <iostream>
 
-typedef size_t TIME;
+typedef float TIME;
